Refactor: Move Runner to agent/ package
Based in part on changes by laszlocph (kube branch):
7f4a0c8c12845a208e53bf11193c7b128e18bfc8 Factor into Runner
f7fe9edd0b90dfe8414bd6b9c7c61c8a3304da6c Using Runner in server
